+++ tools/ui/src/lib/stores/agentic-gates.svelte.ts
@@ -1,0 +1,208 @@
+/**
+ * AgenticGates - User interaction gates for the agentic loop
+ *
+ * Owns the state the loop waits on between turns: tool permission requests,
+ * turn-limit continue prompts and queued steering messages. The loop awaits
+ * requestPermission/requestContinue; the UI resolves them through
+ * resolvePermission/resolveContinue. Owned by agenticStore, no host coupling.
+ */
+
+import { ToolPermissionDecision } from '$lib/enums';
+// direct imports between stores, not via the barrel, to avoid circular deps
+import { permissionsStore } from '$lib/stores/permissions.svelte';
+import { toolsStore } from '$lib/stores/tools.svelte';
+import type { DatabaseMessageExtra, SteeringMessage } from '$lib/types';
+import { SvelteMap } from 'svelte/reactivity';
+
+export class AgenticGates {
+	/** Dedicated reactive state for pending permission requests (ensures immediate UI updates) */
+	private _pendingPermissions = new SvelteMap<
+		string,
+		{ toolName: string; serverLabel: string } | null
+	>();
+	/** Resolve functions for pending permission Promises; nothing derives from this map */
+	private _permissionResolvers = new SvelteMap<string, (decision: ToolPermissionDecision) => void>();
+
+	/** Dedicated reactive state for pending continue requests (turn limit reached) */
+	private _pendingContinueRequests = new SvelteMap<string, boolean>();
+	/** Resolve functions for pending continue Promises; nothing derives from this map */
+	private _continueResolvers = new SvelteMap<string, (shouldContinue: boolean) => void>();
+
+	/** Reactive: queued steering messages to inject between turns */
+	private _steeringMessages = new SvelteMap<string, SteeringMessage>();
+
+	pendingPermissionRequest(
+		conversationId: string
+	): { toolName: string; serverLabel: string } | null {
+		return this._pendingPermissions.get(conversationId) ?? null;
+	}
+
+	pendingContinueRequest(conversationId: string): boolean {
+		return this._pendingContinueRequests.get(conversationId) ?? false;
+	}
+
+	resolveContinue(conversationId: string, shouldContinue: boolean): void {
+		const resolver = this._continueResolvers.get(conversationId);
+
+		if (resolver) {
+			this._continueResolvers.delete(conversationId);
+			resolver(shouldContinue);
+		}
+	}
+
+	resolvePermission(conversationId: string, decision: ToolPermissionDecision): void {
+		const resolver = this._permissionResolvers.get(conversationId);
+
+		if (resolver) {
+			this._permissionResolvers.delete(conversationId);
+			resolver(decision);
+		}
+	}
+
+	hasPendingSteeringMessage(conversationId: string): boolean {
+		return this._steeringMessages.has(conversationId);
+	}
+
+	pendingSteeringMessageContent(conversationId: string): string | null {
+		return this._steeringMessages.get(conversationId)?.content ?? null;
+	}
+
+	pendingSteeringMessageExtras(conversationId: string): DatabaseMessageExtra[] | undefined {
+		return this._steeringMessages.get(conversationId)?.extras;
+	}
+
+	/**
+	 * Queue a steering message. When the current agentic turn completes,
+	 * the flow exits and the caller re-sends the message as a normal chat message.
+	 */
+	injectSteeringMessage(
+		conversationId: string,
+		content: string,
+		extras?: DatabaseMessageExtra[]
+	): void {
+		this._steeringMessages.set(conversationId, { content, extras });
+	}
+
+	/**
+	 * Clear the pending steering message without consuming it.
+	 */
+	clearSteeringMessage(conversationId: string): void {
+		this._steeringMessages.delete(conversationId);
+	}
+
+	/**
+	 * Consume and return the pending steering message for re-sending.
+	 * Called by chatStore after the agentic flow exits.
+	 */
+	consumePendingSteeringMessage(conversationId: string): SteeringMessage | null {
+		const msg = this._steeringMessages.get(conversationId);
+
+		if (!msg) return null;
+
+		this._steeringMessages.delete(conversationId);
+
+		return msg;
+	}
+
+	/**
+	 * Drop all pending gate state for a conversation, e.g. when a flow exits.
+	 */
+	clear(conversationId: string): void {
+		this._pendingPermissions.set(conversationId, null);
+		this._permissionResolvers.delete(conversationId);
+		this._pendingContinueRequests.set(conversationId, false);
+		this._continueResolvers.delete(conversationId);
+		this._steeringMessages.delete(conversationId);
+	}
+
+	async requestPermission(
+		conversationId: string,
+		toolName: string,
+		serverLabel: string,
+		signal?: AbortSignal
+	): Promise<ToolPermissionDecision> {
+		const permissionKey = toolsStore.getPermissionKey(toolName);
+
+		if (permissionKey && permissionsStore.hasTool(permissionKey)) {
+			return ToolPermissionDecision.ONCE;
+		}
+
+		this._pendingPermissions.set(conversationId, { serverLabel, toolName });
+
+		return new Promise<ToolPermissionDecision>((resolve) => {
+			if (signal?.aborted) {
+				this._pendingPermissions.set(conversationId, null);
+				resolve(ToolPermissionDecision.DENY);
+
+				return;
+			}
+
+			this._permissionResolvers.set(conversationId, (decision) => {
+				this._pendingPermissions.set(conversationId, null);
+
+				if (decision === ToolPermissionDecision.ALWAYS && permissionKey) {
+					permissionsStore.allowTool(permissionKey);
+				} else if (decision === ToolPermissionDecision.ALWAYS_SERVER) {
+					const serverToolKeys = toolsStore.allTools
+						.filter((t) =>
+							t.serverName
+								? t.serverName === serverLabel
+								: toolsStore.getToolServerLabel(t.definition.function.name) === serverLabel
+						)
+						.map((t) => toolsStore.getPermissionKey(t.definition.function.name)!)
+						.filter((k): k is string => k !== null);
+
+					permissionsStore.allowTools(serverToolKeys);
+				}
+
+				resolve(decision);
+			});
+
+			signal?.addEventListener(
+				'abort',
+				() => {
+					const resolver = this._permissionResolvers.get(conversationId);
+
+					if (resolver) {
+						this._permissionResolvers.delete(conversationId);
+						this._pendingPermissions.set(conversationId, null);
+						resolve(ToolPermissionDecision.DENY);
+					}
+				},
+				{ once: true }
+			);
+		});
+	}
+
+	async requestContinue(conversationId: string, signal?: AbortSignal): Promise<boolean> {
+		this._pendingContinueRequests.set(conversationId, true);
+
+		return new Promise<boolean>((resolve) => {
+			if (signal?.aborted) {
+				this._pendingContinueRequests.set(conversationId, false);
+				resolve(false);
+
+				return;
+			}
+
+			this._continueResolvers.set(conversationId, (shouldContinue) => {
+				this._pendingContinueRequests.set(conversationId, false);
+				resolve(shouldContinue);
+			});
+
+			signal?.addEventListener(
+				'abort',
+				() => {
+					const resolver = this._continueResolvers.get(conversationId);
+
+					if (resolver) {
+						this._continueResolvers.delete(conversationId);
+						this._pendingContinueRequests.set(conversationId, false);
+						resolve(false);
+					}
+				},
+				{ once: true }
+			);
+		});
+	}
+}
